Re-commit

+++ Calculator/src/main/java/com/rajesh/Calculator/CalculatorApplication.java
@@ -23,13 +23,13 @@ public class CalculatorApplication {
 	public int multiply(int num1,int num2) {
 		return num1*num2;
 	}
-	public int divide(int num1,int num2) {
+	public double divide(int num1,int num2) {
 		
 		if(num2==0) {
 			 throw new ArithmeticException("zero is not allowed.."); 
 			
 		}
-		return num1/num2;
+		return  (double) num1/num2;
 	}
